re-organise rule group hooks
